+++ .travis.yml
@@ -2,6 +2,7 @@ language: node_js
 node_js:
   - "8"
 script: "yarn build:ci"
+cache: false
 
 # Send coverage data to Coveralls
 after_success: "cat ./.coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js"
