+++ demo/style.less
@@ -7,11 +7,11 @@ body {
     margin: 15px;
 }
 
-h2 {
+h1 {
     padding: 10px 0;
 }
 
-h4 {
+h2 {
     margin: 5px 0;
 }
