+++ src/components/ebay-combobox/examples/3-nested-options/template.marko
@@ -1,0 +1,13 @@
+<var options=[
+    { value: 1, label: 'Option 1' },
+    { value: 2, label: 'Option 2', selected: true },
+    { value: 3, label: 'Option 3' }
+]/>
+
+<span style="display: inline-block; max-width: 200px;">
+    <ebay-combobox name="formFieldName">
+        <if(true)>
+            <ebay-combobox-option for(option in options) value=option.value label=option.label selected=option.selected/>
+        </if>
+    </ebay-combobox>
+</span>
