+++ src/components/ebay-combobox/examples/4-borderless/template.marko
@@ -1,0 +1,7 @@
+<span style="display: inline-block; max-width: 200px;">
+    <ebay-combobox borderless name="formFieldName">
+        <ebay-combobox-option value="1" label="Option 1"/>
+        <ebay-combobox-option value="2" label="Option 2"/>
+        <ebay-combobox-option value="3" label="Option 3"/>
+    </ebay-combobox>
+</span>
