+++ src/components/ebay-select/examples/2-selected-option/template.marko
@@ -1,7 +1,5 @@
-<span style="display: inline-block; max-width: 200px;">
-    <ebay-select name="formFieldName">
-        <ebay-select-option value="1" label="Option 1"/>
-        <ebay-select-option value="2" label="Option 2" selected/>
-        <ebay-select-option value="3" label="Option 3"/>
-    </ebay-select>
-</span>
+<ebay-select name="formFieldName">
+    <ebay-select-option value="1" label="Option 1"/>
+    <ebay-select-option value="2" label="Option 2" selected/>
+    <ebay-select-option value="3" label="Option 3"/>
+</ebay-select>
