+++ src/components/ebay-select/examples/3-nested-options/template.marko
@@ -4,10 +4,8 @@
     { value: 3, label: 'Option 3' }
 ]/>
 
-<span style="display: inline-block; max-width: 200px;">
-    <ebay-select name="formFieldName">
-        <if(true)>
-            <ebay-select-option for(option in options) value=option.value label=option.label selected=option.selected/>
-        </if>
-    </ebay-select>
-</span>
+<ebay-select name="formFieldName">
+    <if(true)>
+        <ebay-select-option for(option in options) value=option.value label=option.label selected=option.selected/>
+    </if>
+</ebay-select>
