+++ src/components/ebay-select/examples/4-borderless/template.marko
@@ -1,7 +1,5 @@
-<span style="display: inline-block; max-width: 200px;">
-    <ebay-select borderless name="formFieldName">
-        <ebay-select-option value="1" label="Option 1"/>
-        <ebay-select-option value="2" label="Option 2"/>
-        <ebay-select-option value="3" label="Option 3"/>
-    </ebay-select>
-</span>
+<ebay-select borderless name="formFieldName">
+    <ebay-select-option value="1" label="Option 1"/>
+    <ebay-select-option value="2" label="Option 2"/>
+    <ebay-select-option value="3" label="Option 3"/>
+</ebay-select>
