+++ src/components/ebay-select/examples/5-disabled/template.marko
@@ -1,4 +1,4 @@
-<ebay-select name="formFieldName">
+<ebay-select disabled name="formFieldName">
     <ebay-select-option value="1" label="Option 1"/>
     <ebay-select-option value="2" label="Option 2"/>
     <ebay-select-option value="3" label="Option 3"/>
