+++ src/components/ebay-select/index.js
@@ -1,17 +1,10 @@
 const markoWidgets = require('marko-widgets');
-const Expander = require('makeup-expander');
-const scrollKeyPreventer = require('makeup-prevent-scroll-keys');
 const emitAndFire = require('../../common/emit-and-fire');
-const eventUtils = require('../../common/event-utils');
 const processHtmlAttributes = require('../../common/html-attributes');
 const observer = require('../../common/property-observer');
 const template = require('./template.marko');
 
-const comboboxOptionsClass = 'combobox__options';
-const comboboxExpanderClass = 'combobox__control';
-const comboboxHostSelector = `.${comboboxExpanderClass} > input`;
-const comboboxBtnClass = 'combobox__control';
-const comboboxOptionSelector = '.combobox__option[role=option]';
+const selectOptionSelector = 'select > option';
 
 function getInitialState(input) {
     const options = (input.options || []).map(option => {
@@ -22,8 +15,7 @@ function getInitialState(input) {
             value: option.value,
             label: option.label,
             selected: Boolean(selected),
-            htmlAttributes: processHtmlAttributes(option),
-            renderBody: option.renderBody
+            htmlAttributes: processHtmlAttributes(option)
         };
     });
 
@@ -38,146 +30,60 @@ function getInitialState(input) {
         class: input.class,
         options,
         selected: selectedOption,
-        grow: input.grow,
         borderless: Boolean(input.borderless),
+        disabled: Boolean(input.disabled),
         htmlAttributes: processHtmlAttributes(input)
     };
 }
 
 function getTemplateData(state) {
-    const comboboxClass = ['combobox', state.class];
-    const btnClass = [comboboxBtnClass];
-    const optionsClass = [comboboxOptionsClass];
+    const selectClass = ['select', state.class];
 
     if (state.borderless) {
-        btnClass.push('combobox__control--borderless');
+        selectClass.push('select--borderless');
     }
 
     return {
-        class: comboboxClass,
-        btnClass,
-        optionsClass,
+        class: selectClass,
         name: state.name,
         selectedOption: state.selected,
         options: state.options,
-        grow: state.grow,
+        disabled: state.disabled,
         htmlAttributes: state.htmlAttributes
     };
 }
 
 function init() {
-    const optionEls = this.el.querySelectorAll(comboboxOptionSelector);
+    const optionEls = this.el.querySelectorAll(selectOptionSelector);
+
+    const selectedIndexObserverCallback = (selectedIndex) => {
+        if (optionEls[selectedIndex]) {
+            this.processAfterStateChange(optionEls[selectedIndex]);
+        } else {
+            console.error('Please provide a valid index.');
+        }
+    };
+
+    const valueObserverCallback = (optionValue) => {
+        const optionFind = (option) => option.value.toString() === optionValue.toString();
+        const newOptionSelected = this.state.options.find(optionFind);
+        let optionIndex;
+
+        if (newOptionSelected) {
+            this.state.options.map((option, i) => {
+                if (option.value === newOptionSelected.value) {
+                    optionIndex = i;
+                }
+            });
+
+            this.processAfterStateChange(optionEls[optionIndex]);
+        }
+    };
 
     if (this.state.options && this.state.options.length > 0) {
-        this.expander = new Expander(this.el, {
-            autoCollapse: true,
-            expandOnClick: true,
-            contentSelector: `.${comboboxOptionsClass}`,
-            hostSelector: comboboxHostSelector,
-            hostContainerClass: `${comboboxBtnClass}`,
-            simulateSpacebarClick: true
-        });
-
-        observer.observeRoot(this, ['selected'], (index) => {
-            this.processAfterStateChange(optionEls[index]);
-        });
-
-        const selectedObserverCallback = (optionEl) => {
-            this.processAfterStateChange(optionEl);
-        };
-
-        this.optionEls = optionEls.forEach((optionEl, i) => {
-            observer.observeInner(this, optionEl, 'selected', `options[${i}]`, 'options', selectedObserverCallback);
-        });
-
-        scrollKeyPreventer.add(this.el.querySelector(comboboxHostSelector));
-        scrollKeyPreventer.add(this.el.querySelector(`.${comboboxOptionsClass}`));
+        observer.observeRoot(this, ['selectedIndex'], selectedIndexObserverCallback);
+        observer.observeRoot(this, ['value'], valueObserverCallback);
     }
-}
-
-function handleExpand() {
-    emitAndFire(this, 'listbox-expand');
-}
-
-function handleCollapse() {
-    emitAndFire(this, 'listbox-collapse');
-}
-
-/**
- * Handle mouse click for option
- * @param {MouseEvent} event
- */
-function handleOptionClick(event) {
-    let el;
-
-    // find the element with the data
-    // start with the target element
-    if (event.target.dataset.optionValue) {
-        el = event.target;
-    // check up the tree one level (in case option-label or status was clicked)
-    } else if (event.target.parentNode.dataset.optionValue) {
-        el = event.target.parentNode;
-    }
-
-    this.processAfterStateChange(el);
-    this.expander.collapse();
-    this.el.querySelector(comboboxHostSelector).focus();
-}
-
-/**
- * Handle selection of options when the combobox is closed
- * https://ebay.gitbooks.io/mindpatterns/content/input/listbox.html#keyboard
- * @param {KeyboardEvent} event
- */
-function handleComboboxKeyDown(event) {
-    eventUtils.handleUpDownArrowsKeydown(event, () => {
-        const currentSelectedIndex = this.state.options.findIndex(option => option.selected);
-        const options = clearComboboxSelections(this.state.options);
-        const optionEls = this.el.querySelectorAll(comboboxOptionSelector);
-        let selectElementIndex = currentSelectedIndex;
-
-        switch (event.charCode || event.keyCode) {
-            // down
-            case 40:
-                selectElementIndex = traverseOptions(options, currentSelectedIndex, 1);
-                break;
-            // up
-            case 38:
-                selectElementIndex = traverseOptions(options, currentSelectedIndex, -1);
-                break;
-            default:
-                break;
-        }
-
-        options[selectElementIndex].selected = true;
-
-        this.setState('options', options);
-        this.processAfterStateChange(optionEls[selectElementIndex]);
-    });
-
-    eventUtils.handleEscapeKeydown(event, () => {
-        this.expander.collapse();
-        this.el.querySelector(comboboxHostSelector).focus();
-    });
-}
-
-/**
- * Traverse the options forward or backward for the next/prev option
- * @param {Array} options
- * @param {Number} currentIndex
- * @param {Number} distance
- * @returns {Number}
- */
-function traverseOptions(options, currentIndex, distance) {
-    let goToIndex = currentIndex;
-
-    if (distance < 0 && currentIndex !== 0) {
-        goToIndex = (currentIndex + options.length + distance) % options.length;
-    } else if (distance > 0 && currentIndex !== (options.length - 1)) {
-        goToIndex = (currentIndex + distance) % options.length;
-    }
-
-    return goToIndex;
 }
 
 /**
@@ -185,10 +91,10 @@ function traverseOptions(options, currentIndex, distance) {
  * @param {HTMLElement} el
  */
 function processAfterStateChange(el) {
-    const optionValue = el.dataset.optionValue;
+    const optionValue = el.value;
     const optionIndex = Array.prototype.slice.call(el.parentNode.children).indexOf(el);
     this.setSelectedOption(optionValue);
-    emitAndFire(this, 'listbox-change', {
+    emitAndFire(this, 'select-change', {
         index: optionIndex,
         selected: [optionValue],
         el
@@ -200,29 +106,34 @@ function processAfterStateChange(el) {
  * @param {String} optionValue
  */
 function setSelectedOption(optionValue) {
-    const newOptionSelected = this.state.options.filter(option => option.value.toString() === optionValue)[0];
+    const optionFind = (option) => option.value.toString() === optionValue.toString();
+    const newOptionSelected = this.state.options.find(optionFind);
     const newOptionSelectedValue = newOptionSelected && newOptionSelected.value;
-    let options = this.clearComboboxSelections(this.state.options);
+    let options = this.state.options;
 
-    options = options.map(option => {
+    options = options.map((option, i) => {
         if (option.value === newOptionSelectedValue) {
             option.selected = true;
-            this.setState('selected', option);
+            this.setState('selectedIndex', i);
+            this.setState('value', option.value);
+        } else {
+            option.selected = false;
         }
+        this.update();
         return option;
     });
 
     this.setState('options', options);
 }
 
-/**
- * Resets all options to un-selected
- * @param {Array} options
- */
-function clearComboboxSelections(options) {
-    return options.map(option => {
-        option.selected = false;
-        return option;
+function optionChanged(e) {
+    this.setSelectedOption(e.target.value);
+    const optionIndex = e.target.selectedIndex;
+    const optionEls = this.el.querySelectorAll(selectOptionSelector);
+    emitAndFire(this, 'select-change', {
+        index: optionIndex,
+        selected: [e.target.value],
+        el: optionEls[optionIndex]
     });
 }
 
@@ -231,11 +142,7 @@ module.exports = markoWidgets.defineComponent({
     getInitialState,
     getTemplateData,
     init,
-    handleExpand,
-    handleCollapse,
-    handleOptionClick,
-    handleComboboxKeyDown,
     processAfterStateChange,
     setSelectedOption,
-    clearComboboxSelections
+    optionChanged
 });
